bugfix: sentry/worker prematurely killed by master_fd pty exit
As a pty's master_fd was only held at the client, the OS would close it once the client
exited. Closing a pty's master causes any processes for which the slave end
is a controlling terminal to receive a SIGHUP. This would kill (SIGTERM) both the
sentry and the worker, resulting in unclean shitdown.

The bug was caught when I investigated why coverage.py result databases were
occasionally corrupted anr/or incomplete. Turned out those are saved in an
atexit handler, which would get interrupted by the SIGHUP->SIGTERM scenario.